move login fns to separate modules, update bcrypt, add query-string dependency
