+++ nodeAppPostPg/login.ts
@@ -1,17 +1,19 @@
-// const bcrypt = require('bcrypt');
+const bcrypt = require('bcrypt');
+const queryString = require('query-string');
 const routeFns    = require('./routeFunctions.js');
+
 
 import { createToken } from './token';
 
 const Boom = require('boom');
 
-// const hashPassword = (password, cb) => {
-//     bcrypt.genSalt(10, (err, salt)=>{
-//         bcrypt.hash(password, salt, (err, hash)=>{
-//             return cb(err, hash);
-//         });
-//     });
-// }
+const hashPassword = (password, cb) => {
+  bcrypt.genSalt(10, (err, salt) => {
+    bcrypt.hash(password, salt, (err, hash) => {
+      return cb(err, hash);
+    })
+  })
+};
 
 // const handler = (req, res)=>{
 
@@ -42,8 +44,10 @@ const Boom = require('boom');
 // });
 
 const verifyUniqueUser = async (req, res) => {
-
     const payload = req.query;
+
+    const parsed = queryString.parse(req.query);
+
       
     const userInfo = await routeFns.getUsersInternal(req, res, payload);
   
@@ -57,7 +61,58 @@ const verifyUniqueUser = async (req, res) => {
     res(req.payload);
   };
 
-  export { verifyUniqueUser };
+  const user = {
+    email: '123',
+    userName: 'abc',
+    // password: 'xyz',
+    password: '$2a$10$Bt2vRGCw3udVph77lGBx8O1ffXFmEQv7d1gGI35nKzN.C1w.jeD32',
+    admin: false
+  };
+  
+  const verifyCredentials = async (req, res) => {
+    const payload = req.query;
+      
+    const userInfo = await routeFns.getUsersInternal(req, res, payload);
+    // const payload = JSON.parse( req.payload.info);
+  
+    const {password, email, userName} = payload;
+  
+    console.log("pwd", password);
+    console.log("email", email);
+    console.log("userName", userName);
+  
+    // TODO get user from db
+  
+    if (email !== user.email || userName !== user.userName) {
+      return res(Boom.badRequest('invalid credentials'));
+    }
+  
+    bcrypt.compare(password, user.password, (err, isValid) => {
+      if (err) {
+        return res(err);
+      }
+  
+      if (isValid) {
+        res(user);
+      }
+      else {
+        res(Boom.badRequest('user not known'));
+      }
+    });
+  };
+  
+  // const authenticateUserSchema = Joi.alternatives().try(
+  //   Joi.object({
+  //     userName: Joi.string().alphanum().min(2).max(30).required(),
+  //     password: Joi.string().required()
+  //   }),
+  //   Joi.object({
+  //     email: Joi.string().email().required(),
+  //     password: Joi.string().required()
+  //   })
+  // );
+  
+  export { hashPassword, verifyUniqueUser, verifyCredentials };
   
 // export server route
 // module.exports = {
